# Conflicts:
#	BARRACUDA.sh.txt

+++ aegir/tools/system/clear.sh
@@ -120,13 +120,17 @@ if [ -d "/dev/disk" ]; then
   fi
 fi
 
-if [ ! -e "/root/.giant_traffic.cnf" ]; then
+_IF_BCP=$(ps aux | grep '[d]uplicity' | awk '{print $2}')
+
+if [ -z "${_IF_BCP}" ] \
+  && [ ! -e "/var/run/speed_purge.pid" ] \
+  && [ ! -e "/root/.giant_traffic.cnf" ]; then
+  touch /var/run/speed_purge.pid
   echo " " >> /var/log/nginx/speed_purge.log
   echo "speed_purge start `date`" >> /var/log/nginx/speed_purge.log
-  ionice -c2 -n7 -p $$
-  renice 19 -p $$
   nice -n19 ionice -c2 -n7 find /var/lib/nginx/speed/* -mtime +1 -exec rm -rf {} \; &> /dev/null
   echo "speed_purge complete `date`" >> /var/log/nginx/speed_purge.log
+  rm -f /var/run/speed_purge.pid
 fi
 
 touch /var/xdrago/log/clear.done
